ai-kyc-query-assistant-starter/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
google-generativeai==0.8.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
